+++ apps-script/Store.ts
@@ -61,14 +61,14 @@ export class FirebaseStore implements IStore {
     let configCache = configCacheJson ? JSON.parse(configCacheJson) : null;
     if (!configCache) {
       const defaultConfig: Config = {
-        TakeProfit: 0.1,
-        SellAtTakeProfit: true,
-        BuyQuantity: 10,
-        LossLimit: 0.05,
-        SECRET: '',
         KEY: '',
+        SECRET: '',
+        BuyQuantity: 10,
         PriceAsset: StableCoin.USDT,
+        StopLimit: 0.05,
+        ProfitLimit: 0.1,
         SellAtStopLimit: false,
+        SellAtProfitLimit: true,
         SwingTradeEnabled: false,
         PriceProvider: PriceProvider.Binance,
         AveragingDown: false,
@@ -157,20 +157,20 @@ export class FirebaseStore implements IStore {
 }
 
 export type Config = {
-  TakeProfit: number
-  SellAtTakeProfit: boolean
-  BuyQuantity: number
-  LossLimit: number
-  SECRET?: string
   KEY?: string
+  SECRET?: string
   PriceAsset: string
+  BuyQuantity: number
+  StopLimit: number
+  ProfitLimit: number
   SellAtStopLimit: boolean
+  SellAtProfitLimit: boolean
   SwingTradeEnabled: boolean
   PriceProvider: PriceProvider
   /**
    * When averaging down is enabled, all the money gained from selling is used to buy more your existing
    * most unprofitable (in percentage) asset.
-   * If you have assets A (+1% profit), B (-10% loss) and C(-15% loss) and A is sold, the tool will buy more
+   * If you have assets A, B (-10% loss) and C(-15% loss) and A is sold, the tool will buy more
    * of C, and C loss will be averaged down, for example to -7%.
    * Next time, if C turns profitable and is sold, the tool will buy more of B.
    * This way, **if the price decline is temporary** for all of your assets,