+++ CHANGELOG.md
@@ -29,5 +29,5 @@
 # v1.0.0-beta (DEPRECATED)
 
 * Integration with Trading View notifications and Binance to buy or sell coins.
-* A moving stop-loss limit, which goes up together with the price.
+* A moving stop-limit, which goes up together with the price.
 * A simple UI to display currently tracked assets and the total profit counter.